refactored files placing
